Refactor program handling: update ProgramDTO to include userString and adjust related classes

+++ SEmulatorProject/ServerEngine/src/serverengine/logic/engine/EmulatorEngine.java
@@ -322,7 +322,7 @@ public class EmulatorEngine implements Engine {
 
     public void chargeCredits(String amount) {
         try {
-            int creditsToCharge = Integer.parseInt(amount);
+            long creditsToCharge = Long.parseLong(amount);
             if (creditsToCharge < 0) {
                 throw new NumberNotInRangeException(creditsToCharge);
             }
